refactor(crosswalk_traffic_light_estimator): extract core estimation logic (#12273)
* refactor(autoware_crosswalk_traffic_light_estimator): rename node.cpp to crosswalk_traffic_light_estimator.cpp


* refactor(autoware_crosswalk_traffic_light_estimator): extract CrosswalkTrafficLightEstimator logic


* fix(autoware_crosswalk_traffic_light_estimator): value-initialize config_ to fix cppcheck warning


* fix(autoware_crosswalk_traffic_light_estimator): move static member functions to anonymous namespace


* fix(autoware_crosswalk_traffic_light_estimator): add const qualifier to match declaration and definition


---------

(cherry picked from commit 0081336e25170a3c057dfca67ea0c7a50e4f5843)
