+++ perception/autoware_crosswalk_traffic_light_estimator/CMakeLists.txt
@@ -11,6 +11,7 @@ autoware_package()
 include_directories()
 
 ament_auto_add_library(autoware_crosswalk_traffic_light_estimator SHARED
+  src/crosswalk_traffic_light_estimator.cpp
   src/flashing_detection.cpp
   src/node.cpp
 )
